xtest: remove ffa.h from SPMC test to resolve compiler error
commit 5489e94f0566 ("core: ffa: add boot info structs and defines") in
optee_os causes the compiler error below in ffa_spmc_1000.c.

In file included from ffa_spmc_1000.c:6:
ffa.h:196:41: error: implicit declaration of function ‘U’ [-Werror=implicit-function-declaration]
  196 | #define FFA_BOOT_INFO_NAME_LEN          U(16)
      |                                         ^
ffa.h:319:19: note: in expansion of macro ‘FFA_BOOT_INFO_NAME_LEN’
  319 |         char name[FFA_BOOT_INFO_NAME_LEN];
      |                   ^~~~~~~~~~~~~~~~~~~~~~
ffa.h:319:14: error: variably modified ‘name’ at file scope
  319 |         char name[FFA_BOOT_INFO_NAME_LEN];
      |              ^~~~

ffa.h uses the U macro which is only part of OP-TEE's libc and not a
standard macro. This causes a compiler error when compiling a host
application and the libc is provided by the compiler.

The FF-A SPMC test in xtest relies on the arm_ffa_user Linux driver so
it doesn't use any of the definitions of the ffa.h. This makes it safe
to remove the include of this file which ultimately resolves the
compiler error.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>