openssl: remove openssl binaries and header files
Remove the header files and binaries of openssl.  This version of
openssl is not the latest and has security issues.  Instead of updating
this to the latest version, developers should be using the openssl
binaries available on their build system and using
OPTEE_OPENSSL_EXPORT=${STAGING_INCDIR}
for their header files (if needed).

Signed-off-by: Jon Mason <jon.mason@arm.com>