ta: os_test: fix TA time wrap test
In the TA time wrap test, the TA time is set to its maximum value. After
one second delay, the TA time was compared to the system time.

The comparaison with the system time might not be pertinent as it is
most likely more than dozen seconds.

Compare the wrapped time to a fixed value instead.

Signed-off-by: Clement Faure <clement.faure@nxp.com>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>