ta: os_test: undefine TA2TA_BUF_SIZE
TA2TA_BUF_SIZE is used to defined some test arrays. Using a macro
instead of a const int declaration is good as it forces the declared
arrays not to be variable length arrays.

However, TA2TA_BUF_SIZE is defined in the middle of the translation
unit. Because its use is local to ta_entry_ta2ta_memref(), undefine it
after the function to make the visibility of the macro scoped.

Signed-off-by: Vincent Mailhol <mailhol.vincent@wanadoo.fr>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>