xtest: pkcs11_1000.c: Test support of object checksum value computation
This test involves PKCS11_CKA_CHECK_VALUE when enabled, as per the
spec, the attribute can be either the legitimate value recomputed by
the PKCS#11 token or a zero-sized value called a no-value for when
client does not want the attribute to set in an object. This test
invokes Cryptoki API functions C_GenerateKey(), C_CreateObject(),
C_CopyObject(), C_SetAttributeValue(), C_UnwrapKey() and C_DeriveKey()
to perform check value computation and this test query the value using
C_GetAttributeValue().

Signed-off-by: Marouene Boubakri <marouene.boubakri@nxp.com>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>