ta: rename Makefile and create symlink
Rename ta/Makefile to ta/Makefile.gmake and create a symbolic link
Makefile -> Makefile.gmake. The purpose is to work around an issue with
the OP-TEE build system [1] which uses Buildroot and CMake to build
optee_test. While doing so it overwrites ta/Makefile with Buildroot's
own Makefile. This change makes sure the makefile is preserved and the
build system can use $(MAKE) -f path/to/opte_test/ta/Makefile.gmake to
build all the TAs at once instead of manually enumerating them [2] which
is both inelegant and fragile.

Link: https://github.com/OP-TEE/build/ [1]
Link: https://github.com/OP-TEE/build/blob/4.2.0/br-ext/package/optee_test_ext/optee_test_ext.mk#L31-L38 [2]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>