ta: os_test: fix TA time wrap test
In the test for wrapped TA time in test_time() we check that the time
has wrapped and that not too much time has passed. The limit for this is
currently 2 seconds of which 1 second is passed in TEE_Wait(). This is
currently not enough time with the recent introduction of fTPM being
probed in parallel with the test. So in order to avoid occasional false
negatives increase the limit to 60 seconds.

Fixes: a2c1ce3a8c31 ("ta: os_test: fix TA time wrap test")
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>