xtest: pkcs11: Update test_digest for test 1019 to use SHA-512
The GP TEE_AsymmetricSignDigest API requires that when a hash
algorithm is specified, the digestLen must match the length of
the digest for that algorithm. However, for the mechanisms
CKM_ECDSA_SHA384 and CKM_ECDSA_SHA512, the buffer
(digest_test_pattern_sha256) will be shorter.
This change aims to address that discrepancy.