regression_4017: Test few one-by-one byte case in level 0
Change default value of middle_count in regression test 4017 when
test level is below 12. Instead of 0, use 2. This allows to pass
few tests feeding data byte per byte without increasing the
test duration as when test level is 12 or above.

By the way correct the inline comment that describes the "middle_count"
value that is 32 when level >= 12 not the opposite.

Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>