ci: update workflow as was done in optee_os
CI in this optee_test project meets the same disk space issues as in
optee_os. Let's sync the workflow file with the one from optee_os. In
addition to addressing the "no space left on device" issue, it makes
the job more modular in case we want to extend the build to other
configurations.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>